// src/types/themeType.ts
export interface ThemeColors {
    backgroundColor: string;
    backgroundColorMedium: string;
    backgroundColorStrong: string;
    surfaceColor: string;
    surfaceColorSelected: string;
    textColorBase: string;
    surfaceColorDisabled: string;
    surfaceColorHover: string;
    buttonColor: string;
    buttonColorDisabled: string;
    textColorStrong: string;
    textColorMedium: string;
    textColorDisabled: string;
    textColorSelected: string;
    textColorTop: string;
    borderColor: string;
    isDarkTheme: string;
  }
  